+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Namespace keywords differently (see `helins.linux.i2c/default+`)
+
 ### Fixed